FEATURE: Include rejected queued posts in the user archive export
Requested at https://meta.discourse.org/t/where-can-a-user-find-the-post-that-was-rejected-by-the-moderator/165671?u=riking

Field selection was hard to determine here, because there's no guarantee about what's in the json payload. I decided to include the entire thing, as it gets passed to the PostCreator with no field whitelisting.

Additionally, I spotted a stray IP address in the payload field on an older forum.